Adjust zoom and center of preview image.
Use a similar logic to set up SubsamplinbScaleImageView to
what was used previously with ZoomView so that the preview
image matches what the actual Wallpaper will look like.
Also do not allow zooming to sizes that are smaller than
the screen to prevent negative crop sizes (the wallpaper
would be zoomed back anyway).

Bug: 72783685
Change-Id: I9a4fbf361f141b03b953eb92a893945d3c778be8