Avoid NPE crash for unknown category id
While very rare, it could happen that we get an invalid (or null)
category collection id when starting IndividualPickerActivity.
For example, when restoring after save wasn't called, or if the
saved collection id doesn't exist any more (third party wallpaper
uninstalled).
In that case the app is currently crashing due to an NPE. Instead,
we now finish the Activity since there's not much to do without a
Category.

Bug: 79405550
Change-Id: Idfb80a2d8a450392d761b65137eb5ce99cc1a2da