Do not openInputStream twice without closing
StreamableAsset was calling openInputStream and then
calculateRawDimensions (which in turns calls openInputStream)
without closing it. This is problematic when the asset is
being streamed from the micropaper content provider.
There's really no need to open the input stream before
calling calculateRawDimensions as this method will also
check for input stream being null.

Bug: 138404632
Change-Id: I2244ef1a480731480d2a0d1d7c552715a060cd22