[LivePicker 2/n] Add LiveWallpaper preview support
Create a new LivePreviewFragment with the same UI and
functionality of Android's LivePicker.

Note that there's still redundancy between
ImagePreviewFragment and LivePreviewFragment, follow up
CLs will clean this up and share as much as possible.

Still missing is support for delete and settings action
buttons. The new LivePreviewFragment won't be hooked up
until all the functionality is implemented.

Bug: 141391722
Change-Id: I859f9122c368c01575c23f73a80ffdb519ef0e71