Fix the UI of the preview pager broken when only wallpaper tab exists and the display size is largest.
It's because the UI of the preview pager resizes between category tiles page and the wallpaper tiles page. (Not what we expected)
In only wallpaper tab exists case, the category tiles page will not have bottom navigation, but the wallpaper tiles page still has the bottom action bar.
When we are from category tiles page to the wallpaper tiles page, we added a margin bottom to the whole layout(for showing the bottom action bar), so the preview pager is resized.
So now, let the margin bottom only affect the bottom part of the whole layout to avoid the preview pager resizing.

Video: https://drive.google.com/file/d/11926ho5Cup50g3HVjcmaRjsvPApuo2bO/view?usp=sharing

Bug: 159697266
Change-Id: I51a2ec5c57bd2bd568a173e8cf0ac2b63ab66862