Fix the wallpaper offset problem in full screen preview when clicking the small preview to full screen preview.
The small preview shows the wallpaper which the user applied to the system instead of the wallpaper from wallpaper gallery.
This wallpaper has been cropped when applying to the system.
The far left of this cropped wallpaper will be the area the user chooses to apply.
HOME screen and the small preview will show the far left of this cropped wallpaper.
But full screen preview still shows the center of it.
So for this kind of wallpaper, full screen preview should show the far left instead of center.

Video: https://drive.google.com/file/d/129EmLU33d4h3uEqmMDT0UhZmzd3etsN2/view?usp=sharing

Bug: 167651054
Change-Id: Iacde420859dcd81acf15d327b245c5e9924e33c3