Change to import android.app.AlertDialog instead of androidx.appcompat.app.AlertDialog.
There are two kind of AlertDialog(android.app.AlertDialog and androidx.appcompat.app.AlertDialog) in the app.
The UI of androidx.appcompat.app.AlertDialog looks abnormal. Seems because we didn't(can't) apply appcompat theme.
For fixing the abnormal UI of the dialog and being consistent the type of AlertDialog in the app, convert androidx.appcompat.app.AlertDialog to android.app.AlertDialog.

Screenshots:
  Old:
    https://screenshot.googleplex.com/96tKu6mHvyAdA9P.png
    https://screenshot.googleplex.com/9BFYQYs8i4J3qPp.png
    https://screenshot.googleplex.com/BCG2ukSCS6AsgLX.png
    https://screenshot.googleplex.com/4UD5xg6DHa6QFJp.png
  New:
    https://screenshot.googleplex.com/8KMuw8YQZkScmVS.png
    https://screenshot.googleplex.com/6bZAzZZZh7xgDyz.png
    https://screenshot.googleplex.com/6kfUNWe7cLiBpr9.png
    https://screenshot.googleplex.com/8AiemAFErJyr4Wt.png

Bug: 169024462
Change-Id: I0dd005ae708579cdae67425d4d9692743de31e2b