Update WallpaperConnection to reparent WallpaperEngine's SurfaceControl
Instead of scaling and positioning the wallpaper within its surface,
reparent it to the containing SurfaceView.
This lets us get rid of LiveTileOverlay and simplify the code.

CategoryFragment is not working 100% as before, since now we're not
showing LiveWallpaper in the lockscreen preview. But this is only a
temporary state until we get rid of CategoryFragment in the future.

Bug: 179127643
Test: manual

Change-Id: I936ebbad395d806b6935611a6deca9cfef2a12e2