Prevent crashing when mirroring live wallpaper's surface
Make sure the LiveWallpaper thumbnail content provider is closed
properly to avoid hanging to it.

Also don't try to mirror the surface if the destination is null

This doesn't really fix the underlying issue, with the wallpaper
crashing, but it should mitigate it by preventing the picker from
crashing if this happens.

Bug: 185239595
Test: manual
Change-Id: I2cdb2312482470a4f15d660c869ebf07aba49ba4