Ensure wallpaper isn't cropped unnecessarily
This fix is somewhat hard to understand without log lines throughout
the method.

Without this change, availableExtraHeightTopAndBottom seems to always be
-1, which ends up _insetting_ the image 1 pixel at the top and
bottom. Variable names indicate availableExtraHeightTopAndBottom was
expected to _always_ be a positive value, so the image would be _outset_
not inset.

This ends up causing inconsistency between WallpaperPicker's preview and
the wallpaper as applied. ex. on my Pixel 4a, the expected height is
1625, but without rounding, this method ends up returning 1623, slightly
offsetting the image from its expected location.

Bug: 189931209
Test: Log lines throughout ImagePickerFragment, test against a bunch of
wallpapers.

Change-Id: Ic9dc1fe34808c1cc25717369cf49f8b4c1047f7b