[WPP2] Avoid to create view in constructor of floating sheet content
This is because in many cases, we pass the text of titles to the
floating sheet content and they are assigned in the constructor.

The view can not be created before those variables are assigned;
otherwise, we will get empty content in the view.

Bug: 242908637
Test: Before https://drive.google.com/file/d/1ehMO2_hPxlkH04_O7WYktBPbGB7oTusF/view?usp=sharing&resourcekey=0-UCZMg8yCN9Zb9xrs43Td8Q
After https://drive.google.com/file/d/1fHgOCBVEpBUXzWVPcVmaX_rwkAyTW_uT/view?usp=sharing&resourcekey=0-v4Y4Q8UMUsHME45bS7YZLQ

Change-Id: Id9e9c7fb10976deb4aa42f3d3f25c51ea460030e