Don't reset undo system when changing wallpaper.
When the wallpaper is changed due to user selection, the activity is
silently* recreated. This will cause the activity's onCreate to be
called again and, if we had any state in the undo system, for that state
to be cleared by the call to startSession.

This CL fixes that issue.

* When the wallpaper changes, so does the color theme, which causes a
  configuration change that restarts activities. Normally, this can be
  caught with onConfigurationChanged in the activity but, for some
  legacy reason, it does not happen for a theme change.

Bug: 262924643
Test: manually made sure that the RESET button is not removed after
changing wallpapers from the quick switcher and that reset funcitonality
still works after the button is clicked and the confirmation dialog is
confirmed.

Change-Id: I98b50cd86188018fbc900abe03ee3c26399e254f