Selected tab remembered when navigating back.
There is a bug where the selected tab is forgotten when navigating back to the main screen. This is because of the overzealous implementation of the intent-driven way to start the wallpaper picker on the home tab when started from the home screen long press. Each time we navigate back to the main screen, the intent is replayed in the fragment and the view-model is forced to switch back to the home screen, even though we were on the lock screen when we "left" it.

This CL fixes that issue by first checking the saved instance state in
the view-model and trusting that more highly than whatever is passed
into the method call.

Fix: 265824257
Test: unit test case added
Test: manually verified that returning to the activity and/or navigating
back from the Shortcuts screen faithfully remains on the lock screen tab

Change-Id: I3a9b8d07d019072233c3990538065e631d2ae66e