Makes DuoTabs into text views.
Using Button instead of TextView adds an elevation that results in the
undesirable drop shadow seen in the light theme as described in the
attached bug. Changing them to TextView fixes this.

Also, made a few fields private.

Bug: 266252315
Test: manually verified that the navigation tabs at the top of the
wallpaper picker screen look flat (without a drop shadow) on both dark
and light themes.

Change-Id: I5cf1da47294e103cd707a6aa1469043027069d20