UX polish for lock screen preview (2/3).
If shown in the context of the quick affordance picker, the following apply:
1. The wallpaper is dim
2. The clock is dim
3. The unselected quick affordance is dim
4. The selected quick affordance is not dim and is a bit bigger

If shown in the context of the normal lock screen or the preview of the
lock screen on the main page of the wallpaper picker, none of the above
applies.

Test: Manually verified this matches the mocks in quick affordance
picker, wallpaper picker, and the lock screen.
Fix: 266013381
Bug: 266116562

Change-Id: I44140b3f261fd52d2d6e75bf8919ce2d1017b595