Don't return to shortcuts when wallpaper changed.
There's a problem that, when the activity is recreated due to a color or
wallpaper change, if we opened the wallpaper picker from Settings >
Display > Lock screen > Shortcuts, the activity re-navigates back to the
shortcuts screen.

This is caused by the fact that the destination extra is re-read in the
activity. The solution is "consume" that extra by removing it from the
intent.

Fix: 267801935
Test: manually verified that the bug doesn't reproduce anymore. When
changing wallpaper or color from the quick switchers, the activity is
recreated but stays on the main screen.

Change-Id: I8cae15394b045043f71440f1d4ab398fcffd8b85