Fixes duplicative section controller issue.
The CustomizationPickerBinder was instantiating all the section
controllers each time the fragment was returned to, resulting in the
duplicative section controller memory leak issue described in the
attached bug.

This is because the binder repeats the call to the section controller
factory function each time the lifecycle of the fragment goes through
STARTED and becuse the fragment creates the binding each time
onCreateView is called.

The approach is simple: the fragment cancels the previous binding before
creating a new one.

Also fixed an issue where we were not filtering unavailable section
controllers in the revamped UI.

Fix: 269525182
Test: manually verified, using log statements, that we only create one
section controller of each type, each time we return to the main screen
Test: manually verified that a previously-showing unavailable section
controller (clocks on a 6 pro) is no longer present.

Change-Id: I874393dbf9bc48582cd0544cda087bf7286fff8c