Fix inconsistent colors returning after setting wallpaper
Inconsistent colors were being returned after setting wallpaper
partially because recent refactoring separated the scroll containers for
the home and lock screen tabs. However, both the home and lock screen
preview sections are important for loading the initial home and lock
wallpaper colors. To fix the issue, the home and lock sections are
adjusted to always be created when the customization picker fragment is
binded no matter which tab is being displayed.

Bug: 279467745
Test: manually verified, and verified with print statements
Change-Id: Ia4d5394242f8ead659f0fc6b0a43f9e309ca6dc0