Clock text adjustment for large text (1/2)
Set clock text to be max 2 lines, and hyphenate. Also added changes
to re-render the preview when surface view sizing changes, because when
properties to adjust the number of lines of text is added to a TextView,
the layout re-sizes itself, but the preview does not re-render and
creates an incorrect cropping.

Bug: 278665088
Test: Manually verified, see comment
Change-Id: I3bdcc707383729c080879621ed2a5d3403a374c3