Fixed leak and no clock carousel (2/3)
Previously, ClockRegistry is bond to activity lifecycle and it's not stable with unregister call followed by register in a short period of time.

Remove the register/unregister ClockRegistry from ClockRegistryProvider (used in activity), and move to app level singleton. Register ClockRegistry listeners when app started and cleaned up when app ended.

Bug: 285348630
Bug: 285321790
Test: resume/destroy WPP then launch LS tab
Test: rotate foldable inner screen
Test: run profiler check no related leaks
Change-Id: I6e28ff61250e5e74b396377ff32bf0d0beeefffe