Added changes for proper reloading of applied creative wallpapers in quick switcher
This CL adds a key called lastUpdatedTimeStamp in order to distinguish
between generated creative wallpapers so that the preview of the newly
generated wallpaper is correctly fetched. This is needed since the key
of the cache is service name right now which is same for all creative
wallpapers of a particular type.

Bug: b/280305522
Test: Tested by building and installing the picker APK, and trying out
generating different images of AI wallpapers

Change-Id: I7230704b17de7c7f00d133e3093438989cace48d