Added a check to make sure we have valid data before we setup image grid
This CL aims at making sure that we have valid data in the items list
before we go ahead and setup the image grid. This is needed since in a
few cases, it can happen that due to loading of different entries in
items list on a background thread, they are populated at a slightly
later time which causes an exception here.

Bug: b/286243498
Test: Tested by building and installing the picker APK, and checking if
everything works as expected before.

Change-Id: I624639f64677f6d085fdcaf86c12587e5656f05f