Added changes in order to manually add padding for the creative categories page
This CL aims at adding the padding manually for the creative category
individual selector pages because adding it to the broader vertical
recycler view leads to a flicker since the parent recycler view does not
get to know when the child recycler view draws over it.

Bug: b/280306347
Test: Tested by installing the picker on local, and looking at the emoji
and AI Wallpaper tiles as well as the pre-existing categories. Video
attached in bug.

Change-Id: Ib1b68204e122b5d5253838b39781646fc270788d