Disable item animator for items in recycler view in order to make sure there is no in-consistency in recycler view when loading of categories take time
When categories take some time to load, we show a loading indicator,
however, we add & remove the categories in recycler view rapidly as a
result of which sometimes there is a mis-match between the dataset and
the adaptor's state in RecyclerView. Disabling the item animator
resolves the inconsistency issues.

Bug: b/285633095
Test: Tested by manually adding a delay of 5 seconds in the picker and
checking if the crash happens or not.

Change-Id: I21a4681c08a7ae8dc2dc0f75b68d878a31ed09c8