Fixed issues in rendering of category tiles
This CL aims at having a loading indicator (outside of recycler view)
which would appear till the time all categories haven't been loaded
instead of having category tiles appear on the screen as they are
fetched. It also makes changes such that instead of notifying
recycler view for each item inserted, we only notify at the
end once all items have been loaded/added. This ensures that
we are able to determine positions in onBindViewHolder() method
accurately w.r.t items inserted into the category list which was
otherwise buggy before since category list was changing as more
categories were added to it. It also disables recycling for MyPhotos
holder specifically which was causing the thumbnail on top of MyPhotos
block to be not visible properly.

Bug: b/291002837
Bug: b/297261712
Test: Tested by building picker, adding a live wallpaper from playstore,
adding a 3p wallpaper app from playstore and checking if all category
tiles appear as expected, video attached to bug

Change-Id: I9b3d1d844e8c7580f50059e61366177c93e09eab