Remove notifications to recycler view with each addCategory
This CL aims at adding fix for a crash that can happen in case there is
any single category is added and by that time, the creative categories
haven't finished loading. As a result, since position of other
categories are calculated in reference to available creative categories,
it causes a crash in the wallpaper picker. This has been happening
especially in the case where you download a live wallpaper and scroll
back to the main picker.

Bug: b/297223588
Test: Tested by building and installing picker on local, downloading a
live wallpaper and checking it crashes. Also by downloading a 3P
wallpaper and wallpaper app.

Change-Id: I41cde870459bef24ae0f122eb6c0dc84d5ded36d