Display workspace based on display size
Added the WorkspacePreviewBinder to bind the workspace in the
SmallPreviewBinder. Created a simplified surface view callback to
replace the need of WorkspaceSurfaceHolderCallback.

Also added changes to allow home screen workspace to show in the small
screen preview fragment based on display size. The
workspace surface currently only shows on one preview.

Bug: 291761856
Test: manually verified in small screen preview
Change-Id: I99a65e3f59ab470de270053c24457a9a52104df2