Made CreativeCategory page compatible for landscape mode on foldables
This CL aims at making sure that the layout is correctly surfaced for
landscape mode for creative category tiles. Currently, the span count
was getting incorrectly calculated in landscape mode of the tablet due
to which the wallpaper tiles were not getting surfaced accurately.

Bug: b/293833661
Test: Tested by building and installing picker, checking the layout on
handheld device, a foldable and a tablet in both portrait and landscape
modes. Also checked the layout of Emoji and also of other types of
wallpapers. Video attached in bug.

Change-Id: Iab72fd9de37356db2db9ee7c4b8ea2d0275fcb9e