Correct the size of the view pager housing the dual preview views
This cl manually sets the height of view pager housing the dual preview views. The reason a custom view pager is
required is because the standard viewpager and viewpager2 classes are not able to wrap their contents when the
content is rendered after the view pager rendering is completed.

Flag: persist.wm.debug.wallpaper_multi_crop

Bug: 303307678
Test: manual
Change-Id: I947403a365f69668dc70560cc0aa355f8cb8db64