Adjust quick affordance section controller to use base flags (1/3)
The quick affordance section controller currently calls runBlocking when
initializing to query flags and check if the feature is available.
However, fragments and section controllers should be getting System UI
flags using base flags, which queries and caches the flags when a flag
is first needed. In addition, the runBlocking function is suspected to
be the cause of the reported crash within the bug.

In this CL, a function to obtain the quick affordance flag value is
added to base flags.

Bug: 296856318
Bug: 290849013
Test: adjusted unit tests for this change and made sure they still pass,
verified manually that quick affordance section controller remains

Change-Id: I17b83322a0ce66b927928dce1cb61efafe6a3faa