Removed extra empty check on appliedWallpaperId list
This CL aims at removing the extra check on appliedWallpaperId list
which isn't needed since we are already comparing the earlier
appliedWallpaperId with the older ones by the equality check. This was
leading to fetching wallpapers twice and therefore cause a crash as
described in the bug.

Bug: b/291003096
Test: Tested by building and installing picker on felix device and
checking in logcat by adding log statements. Also noticed the crash
isn't happening again.
Flag: NA

Change-Id: Iaa4def3c95c7eb8af7c3a6f7dd2579bd01f0499c