Enable dynamic height for wallpaper preview
The wallpaper preview for the new preview screen was set incorrectly to have a static height.
This breaks the UI alignment and prevents adaptive layout.

This CL makes the height of the preview dynamic. i.e. it will fill the space between the top tool bar
and the tabs buttons near the bottom of the screen. The actual sizing of the preview is left to the
constraint layout.

Flag: ACONFIG com.android.wallpaper.multi_crop_preview_ui_flag DISABLED
Bug: 312739985
Test: manual
Change-Id: I7765a682d6b52694cfcc685f03935a91170f2560