Refactor WallpaperPreviewViewModel
Major work of this CL is to remove the use of
SmallPreviewConfigViewModel and WorkspacePreviewConfigViewModel as a
state that's constantly set and written.

Instead, we transform to the data flow pattern so that conforms with the
recommended architecture. In the binders we are listening to the data
flow of the configs that are needed for rendering the wallpaper preview
and workspace preview.

Test: Manually tested
Bug: 291761856
Flag: ACONFIG com.android.wallpaper.multi_crop_preview_ui_flag DEVELOPMENT
Change-Id: I441ccc8125444faece4ec3a2c1003ef8c9222205